global_average_timeseries - updated header, file locations, allowed to run for 5x5 monthly anomalies and actuals (as per code suite, rather than netCDF fields).
make_5x5_monthly_climatology - new. Adapted from make_pentad_climatology but works on 5x5 single files

make_complete_data_files - updated file locations

make_pentad_climatology - updated file locations, set default "period" to "both". Added do3hr option to text output file.

merge_day_night - updated file locations.  Fixed typo in get_fileroot().

utils - bug in netcdf_write() - increased size of integer used for n_grids and n_obs to int32 to save wrap-around error leading to large negative numbers in netCDF output files (CF IDL long integer issues).

+++ EUSTACE_SST_MAT/utils.py
@@ -360,7 +360,7 @@ def netcdf_write(filename, data, n_grids, n_obs, variables, lats, lons, time, do
     # write number of observations 
 
 
-    nc_var = outfile.createVariable("n_grids", np.dtype('int16'), ('time','latitude','longitude',), zlib = do_zip, fill_value = -1)
+    nc_var = outfile.createVariable("n_grids", np.dtype('int32'), ('time','latitude','longitude',), zlib = do_zip, fill_value = -1)
 
     nc_var.long_name = "Number of grid boxes/days/hours going into this grid box"
     nc_var.units = "1"
@@ -368,14 +368,13 @@ def netcdf_write(filename, data, n_grids, n_obs, variables, lats, lons, time, do
     nc_var.standard_name = "Number of grid boxes/days/hours"
     nc_var[:] = np.ma.masked_where(n_grids <=0, n_grids)
 
-    nc_var = outfile.createVariable("n_obs", np.dtype('int16'), ('time','latitude','longitude',), zlib = do_zip, fill_value = -1)
+    nc_var = outfile.createVariable("n_obs", np.dtype('int32'), ('time','latitude','longitude',), zlib = do_zip, fill_value = -1)
 
     nc_var.long_name = "Number of raw observations going into this grid box"
     nc_var.units = "1"
     nc_var.missing_value = -1
     nc_var.standard_name = "Number of Observations"
     nc_var[:] = np.ma.masked_where(n_obs <= 0, n_obs)
-
 
     # Global Attributes
     # from file
